Split into header and source file Re #23266
